pyproject.toml
[project]
name = "bedplot"
version = "0.1.0"
description = "Bed planner for vegetables and seeds"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.8.1",
    "certifi>=2024.12.14",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.0",
    "crispy-bootstrap5>=2024.10",
    "cryptography>=44.0.0",
    "defusedxml>=0.7.1",
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "django-allauth>=65.3.0",
    "django-crispy-forms>=2.3",
    "django-debug-toolbar>=4.4.6",
    "django-extensions>=3.2.3",
    "gunicorn>=23.0.0",
    "idna>=3.10",
    "neapolitan>=24.8",
    "oauthlib>=3.2.2",
    "packaging>=24.2",
    "psycopg>=3.2.3",
    "psycopg-binary>=3.2.3",
    "pycparser>=2.22",
    "pyjwt>=2.10.1",
    "python3-openid>=3.2.0",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "ruff>=0.8.3",
    "sqlparse>=0.5.3",
    "typing-extensions>=4.12.2",
    "urllib3>=2.2.3",
    "whitenoise>=6.8.2",
    "pytest-env>=1.1.5",
    "svg-py>=1.5.0",
]

[dependency-groups]
dev = [
    "import-linter>=2.1",
    "pre-commit>=4.0.1",
    "pytest-django>=4.9.0",
    "pytest>=8.3.4",
]

[tool.pytest_env]
TESTING = "True"
SECRET_KEY = "testgarden"
DATABASE_URL = "sqlite:///:memory:"
